Rename request_worker.js to request-worker.js
